pyproject.toml
[project]
name = "suica-viewer"
version = "0.1.0"
description = ""
authors = [
    {name = "KIRISHIKI Yudai", email = "yudai.kirishiki@protonmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "nfcpy",
]
packages = [
    {include = "felica_viewer"}
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.poetry.scripts]
suica-viewer = "suica_viewer.cli:main"
suica-viewer-gui = "suica_viewer.gui:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
